Updated with utils and adding it to test_task_reminder
